Modified to give the answer to part 2.
./run.sh RunMixer --dataFile=data.txt

Ran for 5.5 hours, but gave the right answer. See 20230318.txt.

Would be faster to use the modulo (%) operator instead of repeated subtraction to reduce the distance.

Didn't try an optimisations. I quit while I was ahead.

+++ day20/State.java
@@ -4,12 +4,12 @@ class State {
 		State obj = new State(0, 1);
 	}
 
-	public State(int id, int v) {
+	public State(int id, long v) {
 		this.id = id;
 		value = v;
 	}
 
-	final int value;
+	final long value;
 
 	final int id;
